'list'

+++ src/page/view.tsx
@@ -154,7 +154,7 @@ const view = memo(() => {
     const deleteBtn = useCallback((e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
 		e.preventDefault();
 		const password = window.prompt('비밀번호를 입력해주세요.', "");
-		if (data && password == data.data.password) {
+		if (data && !Array.isArray(data.data) && data.data.id && password == data.data.password) {
 			if(window.confirm("정말 삭제하시겠습니까?")) {
 				dispatch(DeleteItem({id: data.data.id})).then((result) => {
 					navigate("/main/list");
@@ -163,12 +163,12 @@ const view = memo(() => {
 		} else {
 			window.alert("비밀번호가 틀렸습니다.");
 		}
-	}, []);
+	}, [data]);
 
     return (
         <Container>
             <Spinner visible={loading} />
-            {data && data.data.id ? (
+            {data && data.data && !Array.isArray(data.data) ? (
                 <div className="contentBox">
                     <div className="contentImg">
                         <img src={data.data.file_path} alt="img" />
